Refactored to include all logic in the entitybehavior configuration classes
