+++ DataAccessClient.EntityFrameworkCore.SqlServer.Tests/Configuration/EntityBehaviors/IdentifiableEntityBehaviorConfigurationTests.cs
@@ -1,0 +1,71 @@
+﻿using System.Linq;
+using DataAccessClient.EntityBehaviors;
+using DataAccessClient.EntityFrameworkCore.SqlServer.Configuration.EntityBehaviors;
+using DataAccessClient.EntityFrameworkCore.SqlServer.Tests.TestModels;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using Microsoft.EntityFrameworkCore.Metadata.Conventions;
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
+using Xunit;
+
+namespace DataAccessClient.EntityFrameworkCore.SqlServer.Tests.Configuration.EntityBehaviors
+{
+    [System.Diagnostics.CodeAnalysis.SuppressMessage("Usage", "EF1001:Internal EF Core API usage.", Justification = "<Pending>")]
+    public class IdentifiableEntityBehaviorConfigurationTests
+    {
+        [Fact]
+        public void ModelBuilder_IsIdentifiable_WhenNotCalled_ItShouldHaveNoConfiguration()
+        {
+            // Act
+            var result = new ModelBuilder(new ConventionSet());
+
+            // Assert
+            Assert.Null(result.Entity<TestEntity>().Metadata.FindPrimaryKey());
+        }
+
+        [Fact]
+        public void ModelBuilder_IsIdentifiable_WhenCalled_ItShouldSetIdentifiableConfiguration()
+        {
+            // Arrange
+            var entityTypeBuilder = new ModelBuilder(new ConventionSet());
+
+            // Act
+            var result = IdentifiableEntityBehaviorConfigurationExtensions.ConfigureEntityBehaviorIIdentifiable<TestEntity, int>(entityTypeBuilder);
+
+            // Assert
+            var primaryKey = result.Entity<TestEntity>().Metadata.FindPrimaryKey().Properties.Single();
+            Assert.Equal(nameof(IIdentifiable<int>.Id), primaryKey.Name);
+            Assert.False(primaryKey.IsNullable);
+        }
+
+        [Fact]
+        public void EntityTypeBuilder_IsIdentifiable_WhenNotCalled_ItShouldHaveNoConfiguration()
+        {
+            // Arrange
+            var entityType = new EntityType(typeof(TestEntity), new Model(new ConventionSet()), ConfigurationSource.Explicit);
+
+            // Act
+            var result = new EntityTypeBuilder<TestEntity>(entityType);
+
+            // Assert
+            Assert.Null(result.Metadata.FindPrimaryKey());
+        }
+
+        [Fact]
+        public void EntityTypeBuilder_IsIdentifiable_WhenCalled_ItShouldSetIdentifiableConfiguration()
+        {
+            // Arrange
+            var entityType = new EntityType(typeof(TestEntity), new Model(new ConventionSet()), ConfigurationSource.Explicit);
+            var entityTypeBuilder = new EntityTypeBuilder<TestEntity>(entityType);
+
+            // Act
+            var result = entityTypeBuilder.IsIdentifiable<TestEntity, int>();
+
+            // Assert
+            var primaryKey = result.Metadata.FindPrimaryKey().Properties.Single();
+            Assert.Equal(nameof(IIdentifiable<int>.Id), primaryKey.Name);
+            Assert.False(primaryKey.IsNullable);
+        }
+    }
+}
