+++ DataAccessClient.EntityFrameworkCore.SqlServer.Tests/Configuration/EntityBehaviors/IntegrationsTests/TenantScopableIntegrationTests.cs
@@ -8,7 +8,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Xunit;
 
-namespace DataAccessClient.EntityFrameworkCore.SqlServer.Tests.EntityBehaviors
+namespace DataAccessClient.EntityFrameworkCore.SqlServer.Tests.Configuration.EntityBehaviors.IntegrationsTests
 {
     public class TenantScopableIntegrationTests : DbContextTestBase
     {
