--- DataAccessClient.EntityFrameworkCore.SqlServer.Tests/ModelBuilderExtensionsTests.cs
@@ -1,314 +1,0 @@
-﻿using System.Linq;
-using DataAccessClient.EntityBehaviors;
-using DataAccessClient.EntityFrameworkCore.SqlServer.Tests.TestModels;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Conventions;
-using Xunit;
-
-namespace DataAccessClient.EntityFrameworkCore.SqlServer.Tests
-{
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Usage", "EF1001:Internal EF Core API usage.", Justification =
-        "<Pending>")]
-    public class ModelBuilderExtensionsTests
-    {
-        [Fact]
-        public void IsIdentifiable_WhenNotCalled_ItShouldHaveNoConfiguration()
-        {
-            // Act
-            var result = new ModelBuilder(new ConventionSet());
-
-            // Assert
-            Assert.Null(result.Entity<TestEntity>().Metadata.FindPrimaryKey());
-        }
-
-        [Fact]
-        public void IsIdentifiable_WhenCalled_ItShouldSetIdentifiableConfiguration()
-        {
-            // Arrange
-            var entityTypeBuilder = new ModelBuilder(new ConventionSet());
-
-            // Act
-            var result = ModelBuilderExtensions.ConfigureEntityBehaviorIIdentifiable<TestEntity, int>(entityTypeBuilder);
-
-            // Assert
-            var primaryKey = result.Entity<TestEntity>().Metadata.FindPrimaryKey().Properties.Single();
-            Assert.Equal(nameof(IIdentifiable<int>.Id), primaryKey.Name);
-            Assert.False(primaryKey.IsNullable);
-        }
-        
-        [Fact]
-        public void IsSoftDeletable_WhenNotCalled_ItShouldHaveNoConfiguration()
-        {
-            // Act
-            var result = new ModelBuilder(new ConventionSet());
-
-            // Assert
-            Assert.Null(result.Entity<TestEntity>().Metadata.FindProperty(nameof(ISoftDeletable<int>.IsDeleted)));
-            Assert.Null(result.Entity<TestEntity>().Metadata.FindProperty(nameof(ISoftDeletable<int>.DeletedOn)));
-            Assert.Null(result.Entity<TestEntity>().Metadata.FindProperty(nameof(ISoftDeletable<int>.DeletedById)));
-
-            Assert.Null(result.Entity<TestEntity>().Metadata.GetQueryFilter());
-        }
-
-        [Fact]
-        public void IsSoftDeletable_WhenCalled_ItShouldSetSoftDeletableConfiguration()
-        {
-            // Arrange
-            var entityTypeBuilder = new ModelBuilder(new ConventionSet());
-
-            // Act
-            var result = ModelBuilderExtensions.ConfigureEntityBehaviorISoftDeletable<TestEntity, int>(entityTypeBuilder, x=>x.IsDeleted == false);
-
-            // Assert
-            var isDeleted = result.Entity<TestEntity>().Metadata.FindProperty(nameof(ISoftDeletable<int>.IsDeleted));
-            Assert.Equal(nameof(ISoftDeletable<int>.IsDeleted), isDeleted.Name);
-            Assert.False(isDeleted.IsNullable);
-
-
-            var deletedOn = result.Entity<TestEntity>().Metadata.FindProperty(nameof(ISoftDeletable<int>.DeletedOn));
-            Assert.Equal(nameof(ISoftDeletable<int>.DeletedOn), deletedOn.Name);
-            Assert.True(deletedOn.IsNullable);
-
-            var deletedById = result.Entity<TestEntity>().Metadata.FindProperty(nameof(ISoftDeletable<int>.DeletedById));
-            Assert.Equal(nameof(ISoftDeletable<int>.DeletedById), deletedById.Name);
-            Assert.True(deletedById.IsNullable);
-
-            var deletedQueryFilter = result.Entity<TestEntity>().Metadata.GetQueryFilter();
-            Assert.Equal("Param_0 => (Param_0.IsDeleted == False)", deletedQueryFilter.ToString());
-        }
-
-        [Fact]
-        public void IsTenantScopable_WhenNotCalled_ItShouldHaveNoConfiguration()
-        {
-            // Act
-            var result = new ModelBuilder(new ConventionSet());
-
-            // Assert
-            Assert.Null(result.Entity<TestEntity>().Metadata.FindProperty(nameof(ITenantScopable<int>.TenantId)));
-
-            Assert.Null(result.Entity<TestEntity>().Metadata.GetQueryFilter());
-        }
-
-        [Fact]
-        public void IsTenantScopable_WhenCalled_ItShouldSetTenantScopableConfiguration()
-        {
-            // Arrange
-            var entityTypeBuilder = new ModelBuilder(new ConventionSet());
-
-            // Act
-            var result = ModelBuilderExtensions.ConfigureEntityBehaviorITenantScopable<TestEntity, int>(entityTypeBuilder, x=>x.TenantId == 1);
-
-            // Assert
-            var tenantId = result.Entity<TestEntity>().Metadata.FindProperty(nameof(ITenantScopable<int>.TenantId));
-            Assert.Equal(nameof(ITenantScopable<int>.TenantId), tenantId.Name);
-            Assert.False(tenantId.IsNullable);
-
-            var tenantQueryFilter = result.Entity<TestEntity>().Metadata.GetQueryFilter();
-            Assert.Equal("Param_0 => (Param_0.TenantId == 1)", tenantQueryFilter.ToString());
-        }
-
-        [Fact]
-        public void IsLocalizable_WhenNotCalled_ItShouldHaveNoConfiguration()
-        {
-            // Act
-            var result = new ModelBuilder(new ConventionSet());
-
-            // Assert
-            Assert.Null(result.Entity<TestEntityView>().Metadata.FindProperty(nameof(ILocalizable<string>.LocaleId)));
-
-            Assert.Null(result.Entity<TestEntityView>().Metadata.GetQueryFilter());
-        }
-
-        [Fact]
-        public void IsLocalizable_WhenCalled_ItShouldSetLocalizableConfiguration()
-        {
-            // Arrange
-            var entityTypeBuilder = new ModelBuilder(new ConventionSet());
-
-            // Act
-            var result = ModelBuilderExtensions.ConfigureEntityBehaviorILocalizable<TestEntityView, string>(entityTypeBuilder, x => x.LocaleId == "nl-NL");
-
-            // Assert
-            var localeId = result.Entity<TestEntityView>().Metadata.FindProperty(nameof(ILocalizable<int>.LocaleId));
-            Assert.Equal(nameof(ILocalizable<string>.LocaleId), localeId.Name);
-            Assert.False(localeId.IsNullable);
-
-            var localeQueryFilter = result.Entity<TestEntityView>().Metadata.GetQueryFilter();
-            Assert.Equal("Param_0 => (Param_0.LocaleId == \"nl-NL\")", localeQueryFilter.ToString());
-        }
-
-        [Fact]
-        public void IsCreatable_WhenNotCalled_ItShouldHaveNoConfiguration()
-        {
-            // Act
-            var result = new ModelBuilder(new ConventionSet());
-
-            // Assert
-            Assert.Null(result.Entity<TestEntity>().Metadata.FindProperty(nameof(ICreatable<int>.CreatedOn)));
-            Assert.Null(result.Entity<TestEntity>().Metadata.FindProperty(nameof(ICreatable<int>.CreatedById)));
-        }
-
-        [Fact]
-        public void IsCreatable_WhenCalled_ItShouldSetIdentifiableConfiguration()
-        {
-            // Arrange
-            var entityTypeBuilder = new ModelBuilder(new ConventionSet());
-
-            // Act
-            var result = ModelBuilderExtensions.ConfigureEntityBehaviorICreatable<TestEntity, int>(entityTypeBuilder);
-
-            // Assert
-            var createdOn = result.Entity<TestEntity>().Metadata.FindProperty(nameof(ICreatable<int>.CreatedOn));
-            Assert.Equal(nameof(ICreatable<int>.CreatedOn), createdOn.Name);
-            Assert.False(createdOn.IsNullable);
-
-            var createdById = result.Entity<TestEntity>().Metadata.FindProperty(nameof(ICreatable<int>.CreatedById));
-            Assert.Equal(nameof(ICreatable<int>.CreatedById), createdById.Name);
-            Assert.False(createdById.IsNullable);
-        }
-
-        [Fact]
-        public void IsModifiable_WhenNotCalled_ItShouldHaveNoConfiguration()
-        {
-            // Act
-            var result = new ModelBuilder(new ConventionSet());
-
-            // Assert
-            Assert.Null(result.Entity<TestEntity>().Metadata.FindProperty(nameof(IModifiable<int>.ModifiedOn)));
-            Assert.Null(result.Entity<TestEntity>().Metadata.FindProperty(nameof(IModifiable<int>.ModifiedById)));
-        }
-
-        [Fact]
-        public void IsModifiable_WhenCalled_ItShouldSetIdentifiableConfiguration()
-        {
-            // Arrange
-            var entityTypeBuilder = new ModelBuilder(new ConventionSet());
-
-            // Act
-            var result = ModelBuilderExtensions.ConfigureEntityBehaviorIModifiable<TestEntity, int>(entityTypeBuilder);
-
-            // Assert
-            var modifiedOn = result.Entity<TestEntity>().Metadata.FindProperty(nameof(IModifiable<int>.ModifiedOn));
-            Assert.Equal(nameof(IModifiable<int>.ModifiedOn), modifiedOn.Name);
-            Assert.True(modifiedOn.IsNullable);
-
-            var modifiedById = result.Entity<TestEntity>().Metadata.FindProperty(nameof(IModifiable<int>.ModifiedById));
-            Assert.Equal(nameof(IModifiable<int>.ModifiedById), modifiedById.Name);
-            Assert.True(modifiedById.IsNullable);
-        }
-
-        [Fact]
-        public void IsRowVersionable_WhenNotCalled_ItShouldHaveNoConfiguration()
-        {
-            // Act
-            var result = new ModelBuilder(new ConventionSet());
-
-            // Assert
-            Assert.Null(result.Entity<TestEntity>().Metadata.FindProperty(nameof(IRowVersionable.RowVersion)));
-        }
-
-        [Fact]
-        public void IsRowVersionable_WhenCalled_ItShouldSetIdentifiableConfiguration()
-        {
-            // Arrange
-            var entityTypeBuilder = new ModelBuilder(new ConventionSet());
-
-            // Act
-            var result = ModelBuilderExtensions.ConfigureEntityBehaviorIRowVersionable<TestEntity>(entityTypeBuilder);
-
-            // Assert
-            var rowVersion = result.Entity<TestEntity>().Metadata.FindProperty(nameof(IRowVersionable.RowVersion));
-            Assert.Equal(nameof(IRowVersionable.RowVersion), rowVersion.Name);
-            Assert.True(rowVersion.IsNullable);
-            Assert.True(rowVersion.IsConcurrencyToken);
-        }
-
-        [Fact]
-        public void IsTranslatable_WhenNotCalled_ItShouldHaveNoConfiguration()
-        {
-            // Act
-            var result = new ModelBuilder(new ConventionSet());
-
-            // Assert
-            Assert.Null(result.Entity<TestEntity>().Metadata.FindNavigation(nameof(ITranslatable<TestEntityTranslation, int, string>.Translations)));
-            Assert.Null(result.Entity<TestEntityTranslation>().Metadata.FindProperty(nameof(IEntityTranslation<int, string>.TranslatedEntityId)));
-            Assert.Null(result.Entity<TestEntityTranslation>().Metadata.FindProperty(nameof(IEntityTranslation<int, string>.LocaleId)));
-            Assert.Null(result.Entity<TestEntityTranslation>().Metadata.FindPrimaryKey());
-        }
-
-        [Fact]
-        public void IsTranslatable_WhenCalled_ItShouldSetTranslatableConfiguration()
-        {
-            // Arrange
-            var entityTypeBuilder = new ModelBuilder(new ConventionSet());
-
-            // Act
-            var result = ModelBuilderExtensions.ConfigureEntityBehaviorITranslatable<TestEntity, TestEntityTranslation, int, string>(entityTypeBuilder);
-
-            // Assert
-            Assert.NotNull(result.Entity<TestEntity>().Metadata.FindNavigation(nameof(ITranslatable<TestEntityTranslation, int, string>.Translations)));
-            Assert.NotNull(result.Entity<TestEntityTranslation>().Metadata.FindProperty(nameof(IEntityTranslation<int, string>.TranslatedEntityId)));
-            Assert.NotNull(result.Entity<TestEntityTranslation>().Metadata.FindProperty(nameof(IEntityTranslation<int, string>.LocaleId)));
-
-            Assert.Equal(2, result.Entity<TestEntityTranslation>().Metadata.FindPrimaryKey().Properties.Count);
-            Assert.Contains(result.Entity<TestEntityTranslation>().Metadata.FindPrimaryKey().Properties, x =>x.Name == nameof(IEntityTranslation<TestEntity, int, string>.TranslatedEntityId));
-            Assert.Contains(result.Entity<TestEntityTranslation>().Metadata.FindPrimaryKey().Properties, x =>x.Name == nameof(IEntityTranslation<TestEntity, int, string>.LocaleId));
-        }
-
-
-        [Fact]
-        public void ConfigureEntityBehaviorTranslatedProperties_WhenNotCalled_ItShouldHaveNoConfiguration()
-        {
-            // Act
-            var result = new ModelBuilder(new ConventionSet());
-
-            // Assert
-            Assert.Null(result.Entity<TestEntity>().Metadata.FindNavigation(nameof(TestEntity.Name)));
-        }
-
-        [Fact]
-        public void ConfigureEntityBehaviorTranslatedProperties_WhenCalled_ItShouldConfigureTranslatedProperties()
-        {
-            // Arrange
-            var entityTypeBuilder = new ModelBuilder(new ConventionSet());
-
-            // Act
-            var result = ModelBuilderExtensions.ConfigureEntityBehaviorTranslatedProperties<TestEntity>(entityTypeBuilder);
-
-            // Assert
-            Assert.NotNull(result.Entity<TestEntity>().Metadata.FindNavigation(nameof(TestEntity.Name)));
-            Assert.Equal(typeof(TranslatedProperty<string>), result.Entity<TestEntity>().Metadata.FindNavigation(nameof(TestEntity.Name)).ForeignKey.PrincipalToDependent.ClrType);
-        }
-
-        [Fact]
-        public void ConfigureHasUtcDateTimeProperties_WhenNotCalled_ItShouldHaveNoConfiguration()
-        {
-            // Act
-            var result = new ModelBuilder(new ConventionSet());
-
-            // Assert
-            Assert.Null(result.Entity<TestEntity>().Property(nameof(TestEntity.CreatedOn)).Metadata.GetValueConverter());
-            Assert.Null(result.Entity<TestEntity>().Property(nameof(TestEntity.ModifiedOn)).Metadata.GetValueConverter());
-            Assert.Null(result.Entity<TestEntity>().Property(nameof(TestEntity.DeletedOn)).Metadata.GetValueConverter());
-        }
-
-        [Fact]
-        public void ConfigureHasUtcDateTimeProperties_WhenCalled_ItShouldSetValueConverterForDateTimePropertiesConfiguration()
-        {
-            // Arrange
-            var entityTypeBuilder = new ModelBuilder(new ConventionSet());
-
-            // Act
-            var result = ModelBuilderExtensions.ConfigureHasUtcDateTimeProperties<TestEntity>(entityTypeBuilder, new UtcDateTimeValueConverter());
-
-            // Assert
-            Assert.NotNull(result.Entity<TestEntity>().Property(nameof(TestEntity.CreatedOn)).Metadata.GetValueConverter());
-            Assert.NotNull(result.Entity<TestEntity>().Property(nameof(TestEntity.ModifiedOn)).Metadata.GetValueConverter());
-            Assert.NotNull(result.Entity<TestEntity>().Property(nameof(TestEntity.DeletedOn)).Metadata.GetValueConverter());
-
-            Assert.IsType<UtcDateTimeValueConverter>(result.Entity<TestEntity>().Property(nameof(TestEntity.CreatedOn)).Metadata.GetValueConverter());
-            Assert.IsType<UtcDateTimeValueConverter>(result.Entity<TestEntity>().Property(nameof(TestEntity.ModifiedOn)).Metadata.GetValueConverter());
-            Assert.IsType<UtcDateTimeValueConverter>(result.Entity<TestEntity>().Property(nameof(TestEntity.DeletedOn)).Metadata.GetValueConverter());
-        }
-    }
-}
